+++ package_open_source.ps1
@@ -1,6 +1,6 @@
 param(
     [string]$Destination = '',
-    [string]$Version = '0.7.0-preview.1'
+    [string]$Version = '1.0.0'
 )
 
 $ErrorActionPreference = 'Stop'